\name{hgu133plus2PROSITE}
\alias{hgu133plus2PROSITE}
\title{Maps between Manufacturer Identifiers and PROSITE Identifiers}
\description{
   hgu133plus2PROSITE is an R object that provides mappings between
   manufacturer identifiers and PROSITE identifiers.
}

\details{
The bimap interface for PROSITE is defunct. Please use select() interface to
PROSITE identifiers. See ?AnnotationDbi::select for details.
}

\keyword{datasets}
